broke out edit exercise intoparts
